Adding Grantlee TextDocument library from
https://www.grantlee.org/
https://www.gitorious.org/grantlee/pages/Home

His package used CMake and was a bit clunky to interface with the KJPBS build-system due to his expecting Qt to be on the system path (which doesn't work good when you have 7 or 8 different Qt builds to maintain) and it causes many extra system dependencies that I can't have.  I think the only part of it I will need (or want) to use is the TextDocument piece.  While the Template library (and Script library too) might be useful, I don't see it fitting into KJPBS.

So, I made a .pri include and just pulled out the TextDocument set of classes, linking it statically inline with the main code to avoid all of the build issues with getting this to work on Linux, Mac, Windows, Android, etc, etc, concurrently.



git-svn-id: file:///home/dewhisna/Documents/programs/Bible/svndump/purebiblesearch_150731/code_gitcvt/trunk@495 523e89ec-37fb-4b40-b97f-42a194c95123