Finally got the VNC platform target to work!  KJPBS now supports 7 platforms!
Could not get it to work with a static-built Qt Embedded, even with linking the qt-libvncserver into KJPBS.  Maybe it would have worked if I had built Qt, then built qt-libvncserver, then rebuilt Qt with the new static qt-libvncserver added.  But the easy fix, since vncserver itself is a dynamic library that needs to be installed system-wide, is to just build Qt Embedded as dynamic.  The app itself can be static, just not the main Qt and plugins.

Other notes: the CToolTipEdit needs to have a style more like that of the Emscripten build, as it gave a warning about my window attributes not being supported on Qt 4.8.5.  The menus and dialogs seemed to function correctly with normal exec() logic, unlike that of Emscripten.  And of course it isn't sandboxed and can access the local filesystem.  Got it to run pretty well using Remote Desktop Viewer, but never quite got it to run with noVNC, at least not yet.  It would start to connect and then disconnect, regardless of what webserver was running noVNC.

Linking to additional libraries and/or plugins wasn't necessary with the dynamic plugin, so the .pro file didn't need to be modified -- reverted that from last night's experimentation.

Building with libvncserver:
http://libvncserver.sourceforge.net/
http://github.com/connectedtable/qt-libvncserver
http://github.com/kanaka/noVNC


git-svn-id: file:///home/dewhisna/Documents/programs/Bible/svndump/purebiblesearch_150731/code_gitcvt/trunk@831 523e89ec-37fb-4b40-b97f-42a194c95123